move bpr.py layers to losses.py
