+++ deepr/layers/__init__.py
@@ -1,7 +1,6 @@
 # pylint: disable=unused-import,missing-docstring
 
 from deepr.layers.base import Layer, layer
-from deepr.layers.bpr import BPR, MaskedBPR
 from deepr.layers.click_rank import ClickRank
 from deepr.layers.combinators import Sequential, Select, Rename, Parallel, ActiveMode
 from deepr.layers.core import Product, Sum, Concat, Dense, LogicalAnd, DotProduct, Add, Identity, Conv1d, ToFloat
@@ -16,7 +15,7 @@ from deepr.layers.lookup import (
     table_from_file,
     table_from_mapping,
 )
-from deepr.layers.losses import BPRMax, MaskedBPRMax, NegativeSampling, MaskedNegativeSampling
+from deepr.layers.losses import BPR, MaskedBPR, BPRMax, MaskedBPRMax, NegativeSampling, MaskedNegativeSampling
 from deepr.layers.mask import Equal, NotEqual, BooleanMask, BooleanReduceMode, LookAheadMask, PaddingMask
 from deepr.layers.reduce import Average, WeightedAverage
 from deepr.layers.size import IsMinSize
