+++ deepr/layers/losses.py
@@ -1,10 +1,70 @@
-"""Losses Layer"""
+"""Losses Layers"""
 
 import tensorflow as tf
 
 from deepr.layers import base
 from deepr.layers.reduce import WeightedAverage, Average
 from deepr.utils.broadcasting import make_same_shape
+
+
+class BPR(base.Layer):
+    """Vanilla BPR Loss Layer"""
+
+    def __init__(self, **kwargs):
+        super().__init__(n_in=2, n_out=1, **kwargs)
+
+    def forward(self, tensors, mode: str = None):
+        """Forward method of the layer
+
+        Parameters
+        ----------
+        tensors : Tuple[tf.Tensor]
+            - positives: shape = [batch]
+            - negatives: shape = [batch]
+
+        Returns
+        -------
+        tf.Tensor
+            BPR loss
+        """
+        positives, negatives = tensors
+        positives, negatives = make_same_shape([positives, negatives], broadcast=False)
+        losses = -tf.log_sigmoid(positives - negatives)
+        return Average()(losses, mode)
+
+
+class MaskedBPR(base.Layer):
+    """Masked BPR Loss Layer"""
+
+    def __init__(self, **kwargs):
+        super().__init__(n_in=4, n_out=1, **kwargs)
+
+    def forward(self, tensors, mode: str = None):
+        """Forward method of the layer
+
+        Parameters
+        ----------
+        tensors : Tuple[tf.Tensor]
+            - positives : shape = (batch, num_events)
+            - negatives : shape = (batch, num_events, num_negatives)
+            - mask : shape = (batch, num_events, num_negatives)
+            - weights : shape = (batch, num_events)
+
+        Returns
+        -------
+        tf.Tensor
+            BPR loss
+        """
+        positives, negatives, mask, weights = tensors
+        positives, negatives = make_same_shape([positives, negatives], broadcast=False)
+        # One score per negative : (batch, num_events, num_negatives)
+        scores = -tf.log_sigmoid(positives - negatives)
+        # One loss per event, average of scores : (batch, num_events)
+        event_scores = WeightedAverage()((scores, tf.to_float(mask)))
+        # Each event contributes according to its weight
+        event_weights = weights * tf.to_float(tf.reduce_any(mask, axis=-1))
+        event_losses = event_scores * event_weights
+        return tf.div_no_nan(tf.reduce_sum(event_losses), tf.reduce_sum(event_weights))
 
 
 # noinspection PyTypeChecker
@@ -46,7 +106,7 @@ class BPRMax(base.Layer):
         """ Compute softmax"""
         rj_exp = tf.exp(tensor - tf.reduce_max(tensor, axis=2, keepdims=True))
         sum_exp_negative = tf.reduce_sum(tf.multiply(rj_exp, mask), axis=2, keep_dims=True)
-        softmax_scores = tf.divide(rj_exp, (sum_exp_negative + 1e-8))
+        softmax_scores = tf.div_no_nan(rj_exp, sum_exp_negative)
         return softmax_scores
 
 
@@ -75,6 +135,7 @@ class MaskedBPRMax(BPRMax):
             BPR Max loss
         """
         positives, negatives, mask, weights = tensors
+        mask = tf.cast(mask, tf.float32)
         positives, negatives = make_same_shape([positives, negatives], broadcast=False)
         no_sampled_logits = tf.cast(tf.greater_equal(tf.reduce_sum(mask, -1), 0), tf.float32)
         softmax_scores = self._compute_softmax_scores(negatives, mask)
@@ -151,13 +212,14 @@ class MaskedNegativeSampling(base.Layer):
             Negative Sampling  loss
         """
         positives, negatives, mask, weights = tensors
+        mask = tf.cast(mask, tf.float32)
         true_losses = tf.nn.sigmoid_cross_entropy_with_logits(labels=tf.ones_like(positives), logits=positives)
         sampled_losses = tf.nn.sigmoid_cross_entropy_with_logits(labels=tf.zeros_like(negatives), logits=negatives)
         # filter the values that correspond to mask
         negative_number = tf.reduce_sum(mask, -1)
         sampled_losses = tf.multiply(sampled_losses, mask)
         sampled_losses = tf.reduce_sum(sampled_losses, axis=2)
-        sampled_losses = sampled_losses / (negative_number + 1e-8)
+        sampled_losses = tf.div_no_nan(sampled_losses, negative_number)
 
         losses = tf.add(true_losses, sampled_losses)
         # One loss per event, average of scores : (batch, num_events)
